+++ Cargo.toml
@@ -6,10 +6,10 @@ authors = ["Jeff May <jeff.n.may@gmail.com>"]
 [dependencies]
 rocket = "0.3.6"
 rocket_codegen = "0.3.6"
-rocket_contrib = "*"
-diesel = { version = "1.0", features = ["postgres"] }
-r2d2 = "*"
-r2d2-diesel = "*"
-serde_derive = "*"
-serde_json = "*"
-serde = "*"
+rocket_contrib = "0.3.6"
+diesel = { version = "1.1.1", features = ["postgres"] }
+r2d2 = "0.8.2"
+r2d2-diesel = "1.0.0"
+serde_derive = "1.0.27"
+serde_json = "1.0.9"
+serde = "1.0.27"
