+++ src/models.rs
@@ -1,6 +1,6 @@
 
 #[derive(Queryable, Serialize, Deserialize, Debug)]
-pub struct Member{
+pub struct User {
     id: u32,
     email: String,
     name: String,
