+++ src/web/members.rs
@@ -1,30 +1,22 @@
-use rocket;
-
-use rocket_contrib::Json;
-use r2d2::{Pool, PooledConnection};
-use rocket::State;
+use db::DB;
 use diesel::pg::PgConnection;
 use diesel::prelude::*;
+use models::User;
+use r2d2::{Pool, PooledConnection};
+use rocket::{Route, State};
+use rocket_contrib::Json;
 use std::sync::Arc;
-use super::super::DbConn;
-use super::super::models;
-
-
 
 #[get("/hello/<name>/<age>")]
 pub fn hello(name: String, age: u8) -> String {
     format!("Hello, {} year old named {}!", age, name)
 }
 
-
 #[get("/<id>")]
-pub fn find_by_id(db: DbConn, id: u32) -> Option<Json<models::Member>> {
-
+pub fn find_by_id(db: DB, id: u32) -> Option<Json<User>> {
     None
 }
 
-
-
-pub fn routes() -> Vec<rocket::Route> {
+pub fn routes() -> Vec<Route> {
     routes![hello, find_by_id]
 }
